Remove redundant common repository and Cubit
Deleted redundant `common.dart` in `common/data/repositories` and `common/domain/repositories`. Also, removed `common_cubit.dart` from `common/presentation/manager` which were not required anymore. Added a new microservice Dockerfile for notification, generated relevant files for Firebase configuration for macOS. Also, added new global values in common/presentation and files for new implementation of notifications which includes registering, unregistering of device and handlers for foreground and background notifications. Finally, updated the `env.g.dart` file for new environment changes.

The changes are reflecting the shift from common repository having unused functionalities to a more specific implementation for notification handling in the application. With the inclusion of Firebase in macOS, better support and functionalities can be expected. Also, the change includes handling of app's behavior when on foreground and when on background making the notification feature more robust and effective to use. As the env values are also updated, it ensures application is ready to accommodate the changes made wrt notifications handling.